// web/src/types/flow.ts
export enum FlowRouteType {
	Agent = "agent",
	Sub = "sub",
	Tools = "tools",
	Knowledge = "knowledge",
}